Update wrapper.m
I changed the user interface to reflect what we're doing. The S is for the Spin arrangement in the 1D and 2D Ising Model. The T is for the thermodynamic properties of Energy, Magnetization, Magnetic Susceptibility, and Specific Heat. The user can indicate which two options to choose, I thought it would be more clear than saying "lattice" or "magnetization".